Link Miir::MLIRMIOpen to MIOpen instead of MIOpenDriver
This aligns with the old `find_library(LIBMLIRMIOPEN)` method, which
sets the variable `LIBMLIRMIOPEN` if the mlirmiopen static library is
found. And if so, link MIOpen to LIBMLIRMIOPEN.
Later MIOpen is linked to MIOpenDriver, which is further linked to
other targets, such as fin.

Also update the commit id of mlir to the latest.